Only abort client on SYN-RST if no previous SYN-ACK
When initiating the connection, the client sends an initial SYN
packet. If it has not recieved an answer within 50ms, it sends another
SYN-packet, and thereafter it repeats every 500ms until it recieves an
answer. This can lead to multiple SYN-packets being in-air at any one
time, and the server will only send a SYN-ACK to the first SYN
message, with later ones instead being replied with a SYN-RST to
incidate the server is already busy. Previously, this could sometimes
cause the client to abort if it recieved a SYN-RST after the initial
SYN-ACK, causing it to become hard to run nanoping over connections
with more than 50ms RTT.

Slightly change how the client handles recieved SYN-RST messages, so
that it discards them unless it is waiting for the initial
SYN-ACK. This should still correctly abort clients that attempt to
connect to an already busy server, but should avoid the client from
incorrectly aborting itself after it has successfully astablished a
connection to the server.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>