Fix packet size statistic on server side
The packet size reported by the server at the end of a connection
would be based on the received size of the message. The received size
does not include the padded bytes (from --pad-bytes option) as those
are never read in. This results in an inconsistency in how the packet
size is reported between the client and the server, where the client
reports the size of the sent packets (which will include the padded
bytes) whereas the server will report the size of the received packet
without the padded bytes (and as such it will always just be the size
of the nanoping message).

Address this by making the server report the size of the pong messages
it sends instead of the size of the message in the received
packets. As the --pad-bytes option can be set individually on both the
server and the client, this essentially makes both the client and the
server report the packet sizes they've each been configured to send.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>