Make client send packets at fixed intervals
To send packets periodically, the client relied on a simple sleep
loop, where it would sleep for the specified delay before attempting
to send the next ping packet. As this did not account for the time it
took to send the packets, it results in a small amount of drift.

Rework the logic for how frequently the ping packets are sent to be
similar to IRTT. Make the client attempt to send packets at the start
of every fixed interval window of length --delay us, starting from the
first ping packet. This accounts for all the time spent in the sending
loop, which should counteract any long-term drift.

The example below illustrates the impact of this change when
attempting to send packets at 100ms intervals:

1,t0,1728552771.758872443
2,t0,1728552771.859106761
3,t0,1728552771.959344574
4,t0,1728552772.059549258
5,t0,1728552772.159791972
6,t0,1728552772.260027179
7,t0,1728552772.360247178
...
95,t0,1728552781.193782835
96,t0,1728552781.294181492
97,t0,1728552781.394598095
98,t0,1728552781.495000071
99,t0,1728552781.595349196
100,t0,1728552781.695769267
101,t0,1728552781.796165314

At 100ms intervals, only the first decimal point should change, but
due to the drift one can clearly see that digits past the first
decimal point steadily increasing. As a result of the accumalated
drift, the 101st packet is sent 10.037 seconds after the first, rather
than exactly 10s later as epected (100 * 100ms = 10s).

After this commit, the same setup yeilds the following result:
1,t0,1728552801.807482229
2,t0,1728552801.907819715
3,t0,1728552802.007818464
4,t0,1728552802.107788788
5,t0,1728552802.207819606
6,t0,1728552802.307820577
7,t0,1728552802.407835336
...
95,t0,1728552811.207716435
96,t0,1728552811.307834108
97,t0,1728552811.407822136
98,t0,1728552811.507833074
99,t0,1728552811.607835139
100,t0,1728552811.707794402
101,t0,1728552811.807787534

The digits past the first decimal point are now relatively stable
instead of consistently drifting, and the 101st packet is sent almost
exactly 10 seconds after the first one (10.0003 s).

Note that due to the inherent inaccuracies of using sleep to wait,
even after this fix delays < 1000us (like the default of 100us) will
not be particularly accurate.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>