Add server-side check of src address
Add a check on the server that packets after the connection phase
match the expected client address. If the address differs (i.e. it
comes from a different client) then just ignore the packets unless
they're SYN packets, in which case a SYN-RST reply is sent to indicate
that the server is busy.

As it is now also possible to detect duplicate SYN-packets from the
same client, change behavior to resend a SYN-ACK to the connected
client (instead of a SYN-RST) under the assumption that the initial
SYN-ACK was lost. This makes the process slightly more resilient to
packet loss.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>