Track msg type for emulated TX tstamp
The timestamp index (t0-t3) for emulated TX tstamps was based on if
the nanoping instance was a server or a client. However, with the
introduction of the reverse mode this is no longer sufficent to guess
the correct timestamp index (as the server may now send pings and the
client may echo pongs), and results in incorrect TX timestamp indicies
for emulated timestamps in reverse mode.

Therefore, add the msg type to the emulated TX tstamp messages, so the
timestamp index can be set based on the actual msg type, similar to
how it is done for RX timestamps. Furthermore, only log TX timestamps
for ping and pong messages, similar to RX timestamps are handled.

Overall, this ensures that the timestamp index is always correct, and
that only timestamps for the ping-pong messages are logged.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>