Refactor code for logging timestamps
Move all logic for determining if and how a packet should be
timestamped into log_pkt_tstamp(). This allows implementing all logic
regarding how the packet are timestamped in a single place instead of
spreading it out among the 3 call sites. This also makes future
extensions to the timestamp logging format easier to implement.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>