Log paket direction ("cs" or "sc") in duplex mode
Add functionality to append a direction to the timestamp
index (previously just "t0" - "t3"). Packets going from client to
server are suffixed with "cs" and packets going from server to client
are suffixed with "sc".

Enable this logging of direction when running in duplex mode, so that
the t0 and t1 from the client and servers can be distinguished from
each other.