Improve end report
Make several improvements to the reported statistics shown at the end
of the test, and slightly change format to make it easier to read.

Only report the packet loss in scenarios where it makes sense (client
on forward test or server on reverse test).

Measure the duration of the actual test more accurately, ignoring time
for handshakes and waiting periods (measure time directly in the
client_send_loop() and server_loop() functions).

Report the test duration correctly by ensuring the proper number of
leading zeros after the decimal point are used (would previously
report 9ns as 0.9s instead of 0.000000009s). Get rid of the
timevalprint macros altogether as they're not very flexible and in
practice often results in having to write additional lines of code to
get the desired output.

In addition to the number of TX and RX packets, also report the
average packet rate (calculated from the reported duration). In
addition to the reported number of TX and RX stamps, also report the
fraction of packets with tstamps (calculated from reported number of
TX/RX packets).

Make the test more accurately run for the configured time (if --count
is provided), ignoring time of the initial handshake and waiting
periods (move the alarm call into a later spot in run_client()).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>