Change handling of packet size
Make a number of changes to how the packet size is handled in
nanoping.

The packet size used to be set with the --pad-bytes option, which
would simply specify how many additional bytes (on top of those
required for the nanoping header) should be appended as payload to all
packets. This setting was set independently for the client and the
server (so the client would only be in control of the size of the
packets it sent, and the server would be in control of the packets it
sent). Except the initial SYN-packet, which carries the options as
payload, all packets would be padded with the configured amount,
i.e. both the ping/pong packets as well as other control packets.

Replace the --pad-bytes option with the --ping-size/-s and
--pong-size/-o options. This allows the client to control the size of
both the ping and pong packets, regardless of who sends them. The
server can no longer set any of the packet sizes. Refactor the code
for santitizing client options (sanitize_clientopts() so that the two
new client options can also be handled without it getting too
cumbersome.

Furthermore, specify the total packet size at the IP-layer. So to send
fully sized packets you will now specify 1500 bytes instead of a
padding size of 1456 bytes). Nanoping internally calculates the
required padding to achieve the configured packet sizes. Also make the
final statistics report the packet size in terms of the IP packet size
to keep it consistent.

Additionally, make some changes to the backend that allows the packet
size to be set individually for each transmitted packet, rather than
padding all packets from a nanoping instance to a specific size. Use
this capability to only pad ping and pong packets, leaving all control
packets unpadded.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>