Add ptksize to log
With the introduction of the --probe-schedule option, the packet sizes
may now vary throughout the test. So include the packet size in the
per-packet log output to retain this information. Report packet size
including the IP-header, as done elsewhere in nanoping.

In order to log the correct size of received packets, ensure the full
packet is actually received and not just the nanoping header. At high
packet rate with large packets (i.e. high data rate), this may result
in a considerable amount of extra overhead (much more data needs to be
copied from kernel space to user space).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>